# PERF: Series.to_numpy with float dtype and na_value=np.nan