@topper-123 2023-04-05
Actually, what about `FloatDtype(64|32)`?