@lukemanley 2023-04-05
> Actually, what about `FloatDtype(64|32)`?

EA dtypes dispatch to their own implementation of `to_numpy`. In the case of `FloatDtype(64|32)`, I'm not sure there is anything to be done. `BaseMaskedArray.to_numpy` needs to combine the data array and the mask array when NAs exist.